+++ vars/piperPipelineStageInit.groovy
@@ -101,7 +101,7 @@ void call(Map parameters = [:]) {
     piperStageWrapper (script: script, stageName: stageName, stashContent: [], ordinal: 1, telemetryDisabled: true) {
         def scmInfo = checkout scm
 
-        setupCommonPipelineEnvironment(script: script, customDefaults: parameters.customDefaults,
+        setupCommonPipelineEnvironment(script: script, customDefaults: parameters.customDefaults, scmInfo: scmInfo,
             configFile: parameters.configFile, customDefaultsFromFiles: parameters.customDefaultsFromFiles)
 
         Map config = ConfigurationHelper.newInstance(this)
@@ -145,9 +145,6 @@ void call(Map parameters = [:]) {
         } else {
             initStashConfiguration(script, config.stashSettings, config.verbose?: false)
         }
-
-        setGitUrlsOnCommonPipelineEnvironment(script, scmInfo.GIT_URL)
-        script.commonPipelineEnvironment.setGitCommitId(scmInfo.GIT_COMMIT)
 
         if (config.verbose) {
             echo "piper-lib-os  configuration: ${script.commonPipelineEnvironment.configuration}"
@@ -221,62 +218,6 @@ private void initStashConfiguration (script, stashSettings, verbose) {
     Map stashConfiguration = readYaml(text: libraryResource(stashSettings))
     if (verbose) echo "Stash config: ${stashConfiguration}"
     script.commonPipelineEnvironment.configuration.stageStashes = stashConfiguration
-}
-
-private void setGitUrlsOnCommonPipelineEnvironment(script, String gitUrl) {
-
-    Map url = parseUrl(gitUrl)
-
-    if (url.protocol in ['http', 'https']) {
-        script.commonPipelineEnvironment.setGitSshUrl("git@${url.host}:${url.path}")
-        script.commonPipelineEnvironment.setGitHttpsUrl(gitUrl)
-    } else if (url.protocol in [ null, 'ssh', 'git']) {
-        script.commonPipelineEnvironment.setGitSshUrl(gitUrl)
-        script.commonPipelineEnvironment.setGitHttpsUrl("https://${url.host}/${url.path}")
-    }
-
-    List gitPathParts = url.path.replaceAll('.git', '').split('/')
-    def gitFolder = 'N/A'
-    def gitRepo = 'N/A'
-    switch (gitPathParts.size()) {
-        case 1:
-            gitRepo = gitPathParts[0]
-            break
-        case 2:
-            gitFolder = gitPathParts[0]
-            gitRepo = gitPathParts[1]
-            break
-        case { it > 3 }:
-            gitRepo = gitPathParts[gitPathParts.size()-1]
-            gitPathParts.remove(gitPathParts.size()-1)
-            gitFolder = gitPathParts.join('/')
-            break
-    }
-    script.commonPipelineEnvironment.setGithubOrg(gitFolder)
-    script.commonPipelineEnvironment.setGithubRepo(gitRepo)
-}
-
-/*
- * Returns the parts of an url.
- * Valid keys for the retured map are:
- *   - protocol
- *   - auth
- *   - host
- *   - port
- *   - path
- */
-@NonCPS
-/* private */ Map parseUrl(String url) {
-
-    def urlMatcher = url =~ /^((http|https|git|ssh):\/\/)?((.*)@)?([^:\/]+)(:([\d]*))?(\/?(.*))$/
-
-    return [
-        protocol: urlMatcher[0][2],
-        auth: urlMatcher[0][4],
-        host: urlMatcher[0][5],
-        port: urlMatcher[0][7],
-        path: urlMatcher[0][9],
-    ]
 }
 
 private void setPullRequestStageStepActivation(script, config, List actions) {
